test(runtime): split alarm retry coverage
